Split e2e tests into e2e and validation
(cherry picked from commit 7f6c40cd819821c48ca99ea4f60df7a53db1920a)
