+++ Makefile
@@ -41,6 +41,8 @@ OPERATOR_SDK_VERSION=v1.8.1
 OPM_TOOL_URL=https://api.github.com/repos/operator-framework/operator-registry/releases
 
 TESTS_REPORTS_PATH ?= /tmp/test_e2e_logs/
+VALIDATION_TESTS_REPORTS_PATH ?= /tmp/test_validation_logs/
+
 
 ENVTEST_ASSETS_DIR=$(shell pwd)/testbin
 test: generate fmt vet manifests ## Run unit and integration tests
@@ -49,6 +51,11 @@ test: generate fmt vet manifests ## Run unit and integration tests
 	source ${ENVTEST_ASSETS_DIR}/setup-envtest.sh; fetch_envtest_tools $(ENVTEST_ASSETS_DIR); setup_envtest_env $(ENVTEST_ASSETS_DIR); go test ./... -coverprofile cover.out
 
 all: manager ## Default make target if no options specified
+
+test-validation: generate fmt vet manifests  ## Run validation tests
+	rm -rf ${VALIDATION_TESTS_REPORTS_PATH}
+	mkdir -p ${VALIDATION_TESTS_REPORTS_PATH}
+	USE_LOCAL_RESOURCES=true go test --tags=validationtests -v ./test/validation -ginkgo.v -junit $(VALIDATION_TESTS_REPORTS_PATH) -report $(VALIDATION_TESTS_REPORTS_PATH)
 
 test-e2e: generate fmt vet manifests  ## Run e2e tests
 	rm -rf ${TESTS_REPORTS_PATH}
