+++ test/e2e/e2e.go
@@ -3,14 +3,10 @@ package e2e
 import (
 	"context"
 	"flag"
-	"fmt"
-	"io"
 	"os"
 	"path"
 	"testing"
 	"time"
-
-	ctrl "sigs.k8s.io/controller-runtime"
 
 	. "github.com/onsi/ginkgo"
 	"github.com/onsi/ginkgo/extensions/table"
@@ -19,41 +15,28 @@ import (
 
 	metallbv1alpha1 "github.com/metallb/metallb-operator/api/v1alpha1"
 	metallbv1beta1 "github.com/metallb/metallb-operator/api/v1beta1"
-	"github.com/metallb/metallb-operator/pkg/platform"
 	"github.com/metallb/metallb-operator/pkg/status"
 	"github.com/metallb/metallb-operator/test/consts"
 	testclient "github.com/metallb/metallb-operator/test/e2e/client"
 	"github.com/metallb/metallb-operator/test/e2e/k8sreporter"
+	metallbutils "github.com/metallb/metallb-operator/test/metallb"
 	corev1 "k8s.io/api/core/v1"
-	apiext "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/yaml"
 	goclient "sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-const (
-	// Timeout and Interval settings
-	timeout  = time.Minute * 3
-	interval = time.Second * 2
 )
 
 var autoAssign = false
 
-var TestIsOpenShift = false
 var UseMetallbResourcesFromFile = false
 
-var OperatorNameSpace = "metallb-system"
+var OperatorNameSpace = consts.DefaultOperatorNameSpace
 
 var junitPath *string
 var reportPath *string
 
 func init() {
-	if len(os.Getenv("IS_OPENSHIFT")) != 0 {
-		TestIsOpenShift = true
-	}
-
 	if len(os.Getenv("USE_LOCAL_RESOURCES")) != 0 {
 		UseMetallbResourcesFromFile = true
 	}
@@ -81,66 +64,17 @@ func RunE2ETests(t *testing.T) {
 		rr = append(rr, k8sreporter.New(clients, OperatorNameSpace, *reportPath))
 	}
 
-	RunSpecsWithDefaultAndCustomReporters(t, "Metallb Operator Validation Suite", rr)
+	RunSpecsWithDefaultAndCustomReporters(t, "Metallb Operator E2E Suite", rr)
 }
 
 var _ = Describe("metallb", func() {
-	Context("Platform Check", func() {
-		It("Should have the MetalLB Operator namespace", func() {
-			_, err := testclient.Client.Namespaces().Get(context.Background(), OperatorNameSpace, metav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred(), "Should have the MetalLB Operator namespace")
-		})
-		It("should be either Kubernetes or OpenShift platform", func() {
-			cfg := ctrl.GetConfigOrDie()
-			platforminfo, err := platform.GetPlatformInfo(cfg)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(platforminfo.IsOpenShift()).Should(Equal(TestIsOpenShift))
-		})
-	})
-
-	Context("MetalLB", func() {
-		It("should have the MetalLB Operator deployment in running state", func() {
-			Eventually(func() bool {
-				deploy, err := testclient.Client.Deployments(OperatorNameSpace).Get(context.Background(), consts.MetalLBOperatorDeploymentName, metav1.GetOptions{})
-				if err != nil {
-					return false
-				}
-				return deploy.Status.ReadyReplicas == deploy.Status.Replicas
-			}, timeout, interval).Should(BeTrue())
-
-			pods, err := testclient.Client.Pods(OperatorNameSpace).List(context.Background(), metav1.ListOptions{
-				LabelSelector: fmt.Sprintf("control-plane=%s", consts.MetalLBOperatorDeploymentLabel)})
-			Expect(err).ToNot(HaveOccurred())
-
-			deploy, err := testclient.Client.Deployments(OperatorNameSpace).Get(context.Background(), consts.MetalLBOperatorDeploymentName, metav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred())
-			Expect(len(pods.Items)).To(Equal(int(deploy.Status.Replicas)))
-
-			for _, pod := range pods.Items {
-				Expect(pod.Status.Phase).To(Equal(corev1.PodRunning))
-			}
-		})
-
-		It("should have the MetalLB CRD available in the cluster", func() {
-			crd := &apiext.CustomResourceDefinition{}
-			err := testclient.Client.Get(context.Background(), goclient.ObjectKey{Name: consts.MetalLBOperatorCRDName}, crd)
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		It("should have the MetalLB AddressPool CRD available in the cluster", func() {
-			crd := &apiext.CustomResourceDefinition{}
-			err := testclient.Client.Get(context.Background(), goclient.ObjectKey{Name: consts.MetalLBAddressPoolCRDName}, crd)
-			Expect(err).ToNot(HaveOccurred())
-		})
-	})
-
 	Context("MetalLB deploy", func() {
 		var metallb *metallbv1beta1.MetalLB
 		var metallbCRExisted bool
 
 		BeforeEach(func() {
 			var err error
-			metallb, err = getMetalLB()
+			metallb, err = metallbutils.Get(OperatorNameSpace, UseMetallbResourcesFromFile)
 			Expect(err).ToNot(HaveOccurred())
 			metallbCRExisted = true
 			err = testclient.Client.Get(context.Background(), goclient.ObjectKey{Namespace: metallb.Namespace, Name: metallb.Name}, metallb)
@@ -164,7 +98,7 @@ var _ = Describe("metallb", func() {
 				Expect(daemonset.OwnerReferences).ToNot(BeNil())
 				Expect(daemonset.OwnerReferences[0].Kind).To(Equal("MetalLB"))
 
-				deleteMetalLB(metallb)
+				metallbutils.Delete(metallb)
 			}
 		})
 
@@ -176,7 +110,7 @@ var _ = Describe("metallb", func() {
 						return false
 					}
 					return deploy.Status.ReadyReplicas > 0 && deploy.Status.ReadyReplicas == deploy.Status.Replicas
-				}, timeout, interval).Should(BeTrue())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue())
 
 				pods, err := testclient.Client.Pods(OperatorNameSpace).List(context.Background(), metav1.ListOptions{
 					LabelSelector: "component=controller"})
@@ -198,7 +132,7 @@ var _ = Describe("metallb", func() {
 						return false
 					}
 					return daemonset.Status.DesiredNumberScheduled == daemonset.Status.NumberReady
-				}, timeout, interval).Should(BeTrue())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue())
 
 				pods, err := testclient.Client.Pods(OperatorNameSpace).List(context.Background(), metav1.ListOptions{
 					LabelSelector: "component=speaker"})
@@ -261,7 +195,7 @@ var _ = Describe("metallb", func() {
 			Eventually(func() error {
 				err := testclient.Client.Get(context.Background(), key, addresspool)
 				return err
-			}, timeout, interval).Should(Succeed())
+			}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 			// Checking ConfigMap is created
 			By("By checking ConfigMap is created match the expected configuration")
@@ -271,18 +205,18 @@ var _ = Describe("metallb", func() {
 					return "", err
 				}
 				return configmap.Data[consts.MetalLBConfigMapName], err
-			}, timeout, interval).Should(MatchYAML(expectedConfigMap))
+			}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(expectedConfigMap))
 
 			By("By checking AddressPool resource and ConfigMap are deleted")
 			Eventually(func() bool {
 				err := testclient.Client.Delete(context.Background(), addresspool)
 				return errors.IsNotFound(err)
-			}, timeout, interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
+			}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
 
 			Eventually(func() bool {
 				_, err := testclient.Client.ConfigMaps(OperatorNameSpace).Get(context.Background(), consts.MetalLBConfigMapName, metav1.GetOptions{})
 				return errors.IsNotFound(err)
-			}, timeout, interval).Should(BeTrue())
+			}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue())
 		},
 			table.Entry("Test AddressPool object with default auto assign", "addresspool1", &metallbv1alpha1.AddressPool{
 				ObjectMeta: metav1.ObjectMeta{
@@ -335,14 +269,14 @@ var _ = Describe("metallb", func() {
 			var metallb *metallbv1beta1.MetalLB
 			BeforeEach(func() {
 				var err error
-				metallb, err = getMetalLB()
+				metallb, err = metallbutils.Get(OperatorNameSpace, UseMetallbResourcesFromFile)
 				Expect(err).ToNot(HaveOccurred())
 				metallb.SetName("incorrectname")
 				Expect(testclient.Client.Create(context.Background(), metallb)).Should(Succeed())
 			})
 
 			AfterEach(func() {
-				deleteMetalLB(metallb)
+				metallbutils.Delete(metallb)
 			})
 			It("should not be reconciled", func() {
 				By("checking MetalLB resource status", func() {
@@ -366,19 +300,19 @@ var _ = Describe("metallb", func() {
 			var incorrect_metallb *metallbv1beta1.MetalLB
 			BeforeEach(func() {
 				var err error
-				correct_metallb, err = getMetalLB()
+				correct_metallb, err = metallbutils.Get(OperatorNameSpace, UseMetallbResourcesFromFile)
 				Expect(err).ToNot(HaveOccurred())
 				Expect(testclient.Client.Create(context.Background(), correct_metallb)).Should(Succeed())
 
-				incorrect_metallb, err = getMetalLB()
+				incorrect_metallb, err = metallbutils.Get(OperatorNameSpace, UseMetallbResourcesFromFile)
 				Expect(err).ToNot(HaveOccurred())
 				incorrect_metallb.SetName("incorrectname")
 				Expect(testclient.Client.Create(context.Background(), incorrect_metallb)).Should(Succeed())
 			})
 
 			AfterEach(func() {
-				deleteMetalLB(incorrect_metallb)
-				deleteMetalLB(correct_metallb)
+				metallbutils.Delete(incorrect_metallb)
+				metallbutils.Delete(correct_metallb)
 			})
 			It("should have correct statuses", func() {
 				By("checking MetalLB resource status", func() {
@@ -386,14 +320,14 @@ var _ = Describe("metallb", func() {
 						instance := &metallbv1beta1.MetalLB{}
 						err := testclient.Client.Get(context.TODO(), goclient.ObjectKey{Namespace: incorrect_metallb.Namespace, Name: incorrect_metallb.Name}, instance)
 						Expect(err).ToNot(HaveOccurred())
-						return checkConditionStatus(instance) == status.ConditionDegraded
+						return metallbutils.CheckConditionStatus(instance) == status.ConditionDegraded
 					}, 30*time.Second, 5*time.Second).Should(BeTrue())
 
 					Eventually(func() bool {
 						instance := &metallbv1beta1.MetalLB{}
 						err := testclient.Client.Get(context.TODO(), goclient.ObjectKey{Namespace: correct_metallb.Namespace, Name: correct_metallb.Name}, instance)
 						Expect(err).ToNot(HaveOccurred())
-						return checkConditionStatus(instance) == status.ConditionAvailable
+						return metallbutils.CheckConditionStatus(instance) == status.ConditionAvailable
 					}, 30*time.Second, 5*time.Second).Should(BeTrue())
 
 					// Delete incorrectly named resource
@@ -409,7 +343,7 @@ var _ = Describe("metallb", func() {
 						instance := &metallbv1beta1.MetalLB{}
 						err := testclient.Client.Get(context.TODO(), goclient.ObjectKey{Namespace: correct_metallb.Namespace, Name: correct_metallb.Name}, instance)
 						Expect(err).ToNot(HaveOccurred())
-						return checkConditionStatus(instance) == status.ConditionAvailable
+						return metallbutils.CheckConditionStatus(instance) == status.ConditionAvailable
 					}, 30*time.Second, 5*time.Second).Should(BeTrue())
 				})
 			})
@@ -443,7 +377,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() error {
 					err := testclient.Client.Get(context.Background(), key, addresspool)
 					return err
-				}, timeout, interval).Should(Succeed())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 				// Checking ConfigMap is created
 				By("By checking ConfigMap is created and matches addresspool1 configuration")
@@ -453,7 +387,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`address-pools:
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`address-pools:
 - name: addresspool1
   protocol: layer2
   addresses:
@@ -492,7 +426,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() error {
 					err := testclient.Client.Get(context.Background(), key, addresspool)
 					return err
-				}, timeout, interval).Should(Succeed())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 				// Checking ConfigMap is created
 				By("By checking ConfigMap is created and matches addresspool2 configuration")
@@ -502,7 +436,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`address-pools:
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`address-pools:
 - name: addresspool1
   protocol: layer2
   addresses:
@@ -538,7 +472,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() bool {
 					err := testclient.Client.Delete(context.Background(), addresspool)
 					return errors.IsNotFound(err)
-				}, timeout, interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
+				}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
 
 				By("By checking ConfigMap matches the expected configuration")
 				Eventually(func() (string, error) {
@@ -551,7 +485,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`address-pools:
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`address-pools:
 - name: addresspool2
   protocol: layer2
   auto-assign: false
@@ -581,7 +515,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() bool {
 					err := testclient.Client.Delete(context.Background(), addresspool)
 					return errors.IsNotFound(err)
-				}, timeout, interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
+				}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
 
 				By("By checking ConfigMap matches the expected configuration")
 				Eventually(func() (string, error) {
@@ -594,7 +528,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`
 `))
 
 			})
@@ -604,7 +538,7 @@ var _ = Describe("metallb", func() {
 			Eventually(func() bool {
 				_, err := testclient.Client.ConfigMaps(OperatorNameSpace).Get(context.Background(), consts.MetalLBConfigMapName, metav1.GetOptions{})
 				return errors.IsNotFound(err)
-			}, timeout, interval).Should(BeTrue())
+			}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue())
 		})
 	})
 
@@ -636,7 +570,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() error {
 					err := testclient.Client.Get(context.Background(), key, addresspool)
 					return err
-				}, timeout, interval).Should(Succeed())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 				// Checking ConfigMap is created
 				By("By checking ConfigMap is created and matches addresspool configuration")
@@ -646,7 +580,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`address-pools:
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`address-pools:
 - name: addresspool1
   protocol: layer2
   addresses:
@@ -667,7 +601,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() error {
 					err := testclient.Client.Get(context.Background(), key, addresspool)
 					return err
-				}, timeout, interval).Should(Succeed())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 				addresspool.Spec = metallbv1alpha1.AddressPoolSpec{
 					Protocol: "layer2",
@@ -681,7 +615,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() error {
 					err := testclient.Client.Update(context.Background(), addresspool)
 					return err
-				}, timeout, interval).Should(Succeed())
+				}, metallbutils.Timeout, metallbutils.Interval).Should(Succeed())
 
 				// Checking ConfigMap is updated
 				By("By checking ConfigMap is created and matches updated configuration")
@@ -691,7 +625,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`address-pools:
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`address-pools:
 - name: addresspool1
   protocol: layer2
   auto-assign: false
@@ -720,7 +654,7 @@ var _ = Describe("metallb", func() {
 				Eventually(func() bool {
 					err := testclient.Client.Delete(context.Background(), addresspool)
 					return errors.IsNotFound(err)
-				}, timeout, interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
+				}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue(), "Failed to delete AddressPool custom resource")
 
 				By("Checking ConfigMap is deleted")
 				Eventually(func() (string, error) {
@@ -733,7 +667,7 @@ var _ = Describe("metallb", func() {
 						return "", err
 					}
 					return configmap.Data[consts.MetalLBConfigMapName], err
-				}, timeout, interval).Should(MatchYAML(`
+				}, metallbutils.Timeout, metallbutils.Interval).Should(MatchYAML(`
 `))
 
 			})
@@ -743,90 +677,7 @@ var _ = Describe("metallb", func() {
 			Eventually(func() bool {
 				_, err := testclient.Client.ConfigMaps(OperatorNameSpace).Get(context.Background(), consts.MetalLBConfigMapName, metav1.GetOptions{})
 				return errors.IsNotFound(err)
-			}, timeout, interval).Should(BeTrue())
+			}, metallbutils.Timeout, metallbutils.Interval).Should(BeTrue())
 		})
 	})
 })
-
-func checkConditionStatus(instance *metallbv1beta1.MetalLB) string {
-	availableStatus := false
-	degradedStatus := false
-	for _, condition := range instance.Status.Conditions {
-		if condition.Type == status.ConditionDegraded && condition.Status == metav1.ConditionTrue {
-			degradedStatus = true
-		}
-		if condition.Type == status.ConditionAvailable && condition.Status == metav1.ConditionTrue {
-			availableStatus = true
-		}
-	}
-	if availableStatus && !degradedStatus {
-		return status.ConditionAvailable
-	}
-	if !availableStatus && degradedStatus {
-		return status.ConditionDegraded
-	}
-	return ""
-}
-
-func decodeYAML(r io.Reader, obj interface{}) error {
-	decoder := yaml.NewYAMLToJSONDecoder(r)
-	return decoder.Decode(obj)
-}
-
-func loadMetalLBFromFile(metallb *metallbv1beta1.MetalLB, fileName string) error {
-	f, err := os.Open(fmt.Sprintf("../../config/samples/%s", fileName))
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	return decodeYAML(f, metallb)
-}
-
-// Delete and check the MetalLB custom resource is deleted to avoid status leak in between tests.
-func deleteMetalLB(metallb *metallbv1beta1.MetalLB) {
-	err := testclient.Client.Delete(context.Background(), metallb)
-	if errors.IsNotFound(err) { // Ignore err, could be already deleted.
-		return
-	}
-	Expect(err).ToNot(HaveOccurred())
-
-	Eventually(func() bool {
-		err := testclient.Client.Get(context.Background(), goclient.ObjectKey{Namespace: metallb.Namespace, Name: metallb.Name}, metallb)
-		return errors.IsNotFound(err)
-	}, 1*time.Minute, 5*time.Second).Should(BeTrue(), "Failed to delete MetalLB custom resource")
-
-	Eventually(func() bool {
-		_, err := testclient.Client.Deployments(metallb.Namespace).Get(context.Background(), consts.MetalLBDeploymentName, metav1.GetOptions{})
-		return errors.IsNotFound(err)
-	}, timeout, interval).Should(BeTrue())
-
-	Eventually(func() bool {
-		_, err := testclient.Client.DaemonSets(metallb.Namespace).Get(context.Background(), consts.MetalLBDaemonsetName, metav1.GetOptions{})
-		return errors.IsNotFound(err)
-	}, timeout, interval).Should(BeTrue())
-
-	Eventually(func() bool {
-		pods, _ := testclient.Client.Pods(metallb.Namespace).List(context.Background(), metav1.ListOptions{
-			LabelSelector: fmt.Sprintf("component=%s", consts.MetalLBDeploymentName)})
-		return len(pods.Items) == 0
-	}, timeout, interval).Should(BeTrue())
-
-	Eventually(func() bool {
-		pods, _ := testclient.Client.Pods(metallb.Namespace).List(context.Background(), metav1.ListOptions{
-			LabelSelector: fmt.Sprintf("component=%s", consts.MetalLBDaemonsetName)})
-		return len(pods.Items) == 0
-	}, timeout, interval).Should(BeTrue())
-}
-
-func getMetalLB() (*metallbv1beta1.MetalLB, error) {
-	if !UseMetallbResourcesFromFile {
-		return &metallbv1beta1.MetalLB{ObjectMeta: metav1.ObjectMeta{Name: "metallb", Namespace: OperatorNameSpace}}, nil
-	}
-	metallb := &metallbv1beta1.MetalLB{}
-	if err := loadMetalLBFromFile(metallb, consts.MetalLBCRFile); err != nil {
-		return nil, err
-	}
-	metallb.SetNamespace(OperatorNameSpace)
-	return metallb, nil
-}
