+++ test/validation/validation_test.go
@@ -1,0 +1,11 @@
+// +build validationtests
+
+package validation
+
+import (
+	"testing"
+)
+
+func TestValidation(t *testing.T) {
+	RunValidationTests(t)
+}
